Implemented parallel bibles
